feat: update and refactor location of technical reports
